move abi-loader to subdir
